Enforce shellscript formatting via prettier plugin
Removing shellcheck altogether, it isn't about formatting in the first
place. But also, from experience, what we run (optionally) in the git
hooks should match exactly what is being tested in CI, due to principle
of least surprise.

Furthermore, the shellcheck hook wasn't correctly working: while we were
passing it only the staged files, in case of partially staging changes,
shellcheck would complain about the entire file as it is, not just the
stage (Talisman's pre-commit hook gets this right btw). Not the end of
the world, but it can become annoying/another unexpected surprise.