Change talisman git hook to match ci
Because talisman's pre-push and pre-commit hooks work quite differently
(the latter will only scan the changed lines of commits) we want to run
the same for the hook and in ci (pre-push). Otherwise talisman might
complain on CI while not on the developer's machine, and that's a
frustrating feedback loop we want to avoid.