Improve secrets scan pre-push hook
It's only improved because of an issue in Lefthook, where the necessary
information from stdin is not being passed on to the hook runner.

Therefore, when pushing from one local ref to another on the remote,
like

git push origin master:foreign

the range of outgoing commits won't be scanned correctly. At least we
now detect the current branch from/to the push occurs, which should be
ok for the majority of cases (e.g. `git push`).

We need to wait for https://github.com/evilmartians/lefthook/issues/147
to be resolved.