Fix package import cycle detection test
The way the slice was defined it was only detecting cycles in the
immediate descendent packages of the parent package to start analysis
from.

As a bonus, the test no longer depends on the package structure as found
in the application template, which is guaranteed to change. The strings
used here won't be catched by refactoring in the IDE though and so there
was a high probability that we would have ended up with a useless test
giving false positives all of the time..