Set up Trivy to perform offline scan
There still seem to be issues when Trivy tries to access the maven
repository api for retrieving additional package data, causing the
scanning to timeout.

As a workaround configuring Trivy for offline scanning.

We might want to revert this later, as offline scanning can result in
less vulnerabilities being discovered.

https://github.com/aquasecurity/trivy/issues/3421